Started extracting CMDIRecordProcessor out of MetadataImporter
