+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/MetadataImporter.java
@@ -5,7 +5,6 @@ import com.google.common.collect.Sets;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
@@ -25,9 +24,7 @@ import org.apache.solr.common.params.SolrParams;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import eu.clarin.cmdi.vlo.CommonUtils;
 import eu.clarin.cmdi.vlo.FieldKey;
-import eu.clarin.cmdi.vlo.FacetConstants;
 import eu.clarin.cmdi.vlo.LanguageCodeUtils;
 import eu.clarin.cmdi.vlo.StringUtils;
 import eu.clarin.cmdi.vlo.config.DataRoot;
@@ -41,7 +38,6 @@ import eu.clarin.cmdi.vlo.importer.normalizer.AvailabilityPostNormalizer;
 import eu.clarin.cmdi.vlo.importer.normalizer.CMDIComponentProfileNamePostNormalizer;
 import eu.clarin.cmdi.vlo.importer.normalizer.ContinentNamePostNormalizer;
 import eu.clarin.cmdi.vlo.importer.normalizer.CountryNamePostNormalizer;
-import eu.clarin.cmdi.vlo.importer.normalizer.FormatPostNormalizer;
 import eu.clarin.cmdi.vlo.importer.normalizer.IdPostNormalizer;
 import eu.clarin.cmdi.vlo.importer.normalizer.LanguageCodePostNormalizer;
 import eu.clarin.cmdi.vlo.importer.normalizer.LanguageNamePostNormalizer;
@@ -125,10 +121,7 @@ public class MetadataImporter {
      */
     protected final Map<String, AbstractPostNormalizer> postProcessors;
 
-    /**
-     * Contains MDSelflinks (usually). Just to know what we have already done.
-     */
-    protected final Set<String> processedIds = Sets.newConcurrentHashSet();
+
     /**
      * Some caching for solr documents (we are more efficient if we ram a whole
      * bunch to the solr server at once.
@@ -499,169 +492,6 @@ public class MetadataImporter {
             }
         }
         return result;
-    }
-
-    /**
-     * Process single CMDI file with CMDIDataProcessor
-     *
-     * @param file CMDI input file
-     * @param dataOrigin
-     * @param resourceStructureGraph null to skip hierarchy processing
-     * @param endpointDescription
-     * @throws SolrServerException
-     * @throws IOException
-     */
-    protected void processCmdi(File file, DataRoot dataOrigin, ResourceStructureGraph resourceStructureGraph, EndpointDescription endpointDescription) throws SolrServerException, IOException {
-        nrOfFilesAnalyzed.incrementAndGet();
-        CMDIData cmdiData = null;
-        try {
-            cmdiData = processor.process(file, resourceStructureGraph);
-            if (!idOk(cmdiData.getId())) {
-                cmdiData.setId(dataOrigin.getOriginName() + "/" + file.getName()); //No id found in the metadata file so making one up based on the file name. Not quaranteed to be unique, but we have to set something.
-                nrOfFilesWithoutId.incrementAndGet();
-            }
-        } catch (Exception e) {
-            LOG.error("error in file: {}", file, e);
-            nrOfFilesWithError.incrementAndGet();
-        }
-        if (cmdiData != null) {
-            if (!cmdiData.hasResources()) {
-                nrOfFilesSkipped.incrementAndGet();
-                LOG.warn("Skipping {}, no resource proxy found", file);
-                return;
-            }
-            
-            assert cmdiData.getId() != null; //idOk check guarantees this
-
-            if (processedIds.add(cmdiData.getId())) {
-                SolrInputDocument solrDocument = cmdiData.getSolrDocument();
-                if (solrDocument != null) {
-                    updateDocument(solrDocument, cmdiData, file, dataOrigin, endpointDescription);
-                    if (resourceStructureGraph != null && resourceStructureGraph.getVertex(cmdiData.getId()) != null) {
-                        resourceStructureGraph.getVertex(cmdiData.getId()).setWasImported(true);
-                    }
-                }
-            } else {
-                nrOfFilesSkipped.incrementAndGet();
-                LOG.warn("Skipping {}, already processed id: {}", file, cmdiData.getId());
-            }
-        }
-    }
-
-    /**
-     * Check id for validness
-     *
-     * @param id
-     * @return true if id is acceptable, false otherwise
-     */
-    protected boolean idOk(String id) {
-        return id != null && !id.trim().isEmpty();
-    }
-
-    /**
-     * Adds some additional information from DataRoot to solrDocument, add
-     * solrDocument to document list, submits list to SolrServer every 1000
-     * files
-     *
-     * @param solrDocument
-     * @param cmdiData
-     * @param file
-     * @param dataOrigin
-     * @param endpointDescription
-     * @throws SolrServerException
-     * @throws IOException
-     */
-    protected void updateDocument(SolrInputDocument solrDocument, CMDIData cmdiData, File file, DataRoot dataOrigin, EndpointDescription endpointDescription) throws SolrServerException,
-            IOException {
-        solrDocument.addField(fieldNameService.getFieldName(FieldKey.DATA_PROVIDER), dataOrigin.getOriginName());
-        solrDocument.addField(fieldNameService.getFieldName(FieldKey.ID), cmdiData.getId());
-        solrDocument.addField(fieldNameService.getFieldName(FieldKey.FILENAME), file.getAbsolutePath());
-
-        // data provided by CLARIN's OAI-PMH harvester
-        if(endpointDescription != null) {
-            if(endpointDescription.getOaiEndpointUrl() != null)
-                solrDocument.addField(fieldNameService.getFieldName(FieldKey.OAI_ENDPOINT_URI), endpointDescription.getOaiEndpointUrl());
-            if(endpointDescription.getNationalProject() != null)
-                solrDocument.addField(fieldNameService.getFieldName(FieldKey.NATIONAL_PROJECT), endpointDescription.getNationalProject());
-            if(endpointDescription.getCentreName() != null)
-                solrDocument.addField(fieldNameService.getFieldName(FieldKey.DATA_PROVIDER_NAME), endpointDescription.getCentreName());
-        }
-
-        String metadataSourceUrl = dataOrigin.getPrefix();
-        metadataSourceUrl += file.getAbsolutePath().substring(dataOrigin.getToStrip().length());
-
-        solrDocument.addField(fieldNameService.getFieldName(FieldKey.COMPLETE_METADATA), metadataSourceUrl);
-
-        // add SearchServices (should be CQL endpoint)
-        for (Resource resource : cmdiData.getSearchResources()) {
-            solrDocument.addField(fieldNameService.getFieldName(FieldKey.SEARCH_SERVICE), resource.getResourceName());
-        }
-
-        // add landing page resource
-        for (Resource resource : cmdiData.getLandingPageResources()) {
-            solrDocument.addField(fieldNameService.getFieldName(FieldKey.LANDINGPAGE), resource.getResourceName());
-        }
-
-        // add search page resource
-        for (Resource resource : cmdiData.getSearchPageResources()) {
-            solrDocument.addField(fieldNameService.getFieldName(FieldKey.SEARCHPAGE), resource.getResourceName());
-        }
-
-        // add timestamp
-        Date dt = new Date();
-        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'");
-        solrDocument.addField(fieldNameService.getFieldName(FieldKey.LAST_SEEN), df.format(dt));
-
-        // set number of days since last import to '0'
-        solrDocument.addField(fieldNameService.getFieldName(FieldKey.DAYS_SINCE_LAST_SEEN), 0);
-
-        // add resource proxys      
-        addResourceData(solrDocument, cmdiData);
-
-        LOG.debug("Adding document for submission to SOLR: {}", file);
-
-        solrBridge.addDocument(solrDocument);
-        if (nrOFDocumentsSent.incrementAndGet() % 250 == 0) {
-            LOG.info("Number of documents sent thus far: {}", nrOFDocumentsSent);
-        }
-    }
-
-    /**
-     * Adds two fields FIELD_FORMAT and FIELD_RESOURCE. The Type can be
-     * specified in the "ResourceType" element of an imdi file or possibly
-     * overwritten by some more specific xpath (as in the LRT cmdi files). So if
-     * a type is overwritten and already in the solrDocument we take that type.
-     *
-     * @param solrDocument
-     * @param cmdiData
-     */
-    protected void addResourceData(SolrInputDocument solrDocument, CMDIData cmdiData) {
-        List<Object> fieldValues = solrDocument.containsKey(fieldNameService.getFieldName(FieldKey.FORMAT)) ? new ArrayList<>(solrDocument
-                .getFieldValues(fieldNameService.getFieldName(FieldKey.FORMAT))) : null;
-        solrDocument.removeField(fieldNameService.getFieldName(FieldKey.FORMAT)); //Remove old values they might be overwritten.
-        List<Resource> resources = cmdiData.getDataResources();
-        for (int i = 0; i < resources.size(); i++) {
-            Resource resource = resources.get(i);
-            String mimeType = resource.getMimeType();
-            if (mimeType == null) {
-                if (fieldValues != null && i < fieldValues.size()) {
-                    mimeType = CommonUtils.normalizeMimeType(fieldValues.get(i).toString());
-                } else {
-                    mimeType = CommonUtils.normalizeMimeType("");
-                }
-            }
-
-            FormatPostNormalizer processor = new FormatPostNormalizer();
-            mimeType = processor.process(mimeType, null).get(0);
-
-            // TODO check should probably be moved into Solr (by using some minimum length filter)
-            if (!mimeType.equals("")) {
-                solrDocument.addField(fieldNameService.getFieldName(FieldKey.FORMAT), mimeType);
-            }
-            solrDocument.addField(fieldNameService.getFieldName(FieldKey.RESOURCE), mimeType + FacetConstants.FIELD_RESOURCE_SPLIT_CHAR
-                    + resource.getResourceName());
-        }
-        solrDocument.addField(fieldNameService.getFieldName(FieldKey.RESOURCE_COUNT), resources.size());
     }
 
     /**
